# app/Models/Team.php
<?php

namespace App\Models;

use App\Enum\Region;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Team extends Model
{
    use HasFactory;

    protected $casts = [
        'region' => Region::class,
    ];
}